+++ src/Commands.cpp
@@ -87,6 +87,7 @@ static SeqStrings gCommandNames =
     "CmdFindToggleMatchCase\0"
     "CmdSaveAnnotations\0"
     "CmdSaveAnnotationsNewFile\0"
+    "CmdPdfBake\0"
     "CmdEditAnnotations\0"
     "CmdDeleteAnnotation\0"
     "CmdZoomFitPage\0"
@@ -298,6 +299,7 @@ static i32 gCommandIds[] = {
     CmdFindToggleMatchCase,
     CmdSaveAnnotations,
     CmdSaveAnnotationsNewFile,
+    CmdPdfBake,
     CmdEditAnnotations,
     CmdDeleteAnnotation,
     CmdZoomFitPage,
@@ -509,6 +511,7 @@ SeqStrings gCommandDescriptions =
     "Find: Toggle Match Case\0"
     "Save Annotations to existing PDF\0"
     "Save Annotations to a new PDF\0"
+    "Bake PDF File\0"
     "Edit Annotations\0"
     "Delete Annotation\0"
     "Zoom: Fit Page\0"